+++ src/main/java/tk/cth451/transitrailmod/blocks/prototype/CustomDirectionBlock.java
@@ -1,0 +1,42 @@
+package tk.cth451.transitrailmod.blocks.prototype;
+
+import java.util.List;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.material.Material;
+import net.minecraft.block.properties.PropertyDirection;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.util.AxisAlignedBB;
+import net.minecraft.util.BlockPos;
+import net.minecraft.util.EnumFacing;
+import net.minecraft.world.World;
+
+public abstract class CustomDirectionBlock extends Block {
+	
+	public static final PropertyDirection FACING = PropertyDirection.create("facing", EnumFacing.Plane.HORIZONTAL);
+
+	public CustomDirectionBlock(Material materialIn) {
+		super(materialIn);
+	}
+	
+	@Override
+	public IBlockState onBlockPlaced(World worldIn, BlockPos pos, EnumFacing facing, float hitX, float hitY, float hitZ, int meta, EntityLivingBase placer){
+		// set facing to the direction player is facing
+		IBlockState state = super.onBlockPlaced(worldIn, pos, facing, hitX, hitY, hitZ, meta, placer);
+		return this.getFacingState(state, placer);
+	}
+	
+	protected IBlockState getFacingState(IBlockState state, EntityLivingBase placer){
+		EnumFacing thisFacing = placer.getHorizontalFacing();
+		return state.withProperty(FACING, thisFacing);
+	}
+	
+	@Override
+	public void addCollisionBoxesToList(World worldIn, BlockPos pos, IBlockState state, AxisAlignedBB mask, List list,
+			Entity collidingEntity) {
+		this.setBlockBoundsBasedOnState(worldIn, pos);
+		super.addCollisionBoxesToList(worldIn, pos, state, mask, list, collidingEntity);
+	}
+}