+++ src/main/java/tk/cth451/transitrailmod/init/ModBlocks.java
@@ -10,11 +10,13 @@ import tk.cth451.transitrailmod.References;
 import tk.cth451.transitrailmod.blocks.ClosedPlatformDoorBlock;
 import tk.cth451.transitrailmod.blocks.ClosedPlatformPanelBlock;
 import tk.cth451.transitrailmod.blocks.ClosedPlatformTop;
+import tk.cth451.transitrailmod.blocks.FluorescentLamp;
 import tk.cth451.transitrailmod.blocks.HungArrowSign;
 import tk.cth451.transitrailmod.blocks.LogoBlock;
 import tk.cth451.transitrailmod.blocks.PlatformArrowSign;
 import tk.cth451.transitrailmod.blocks.PlatformGateBlock;
 import tk.cth451.transitrailmod.blocks.PlatformPanelBlock;
+import tk.cth451.transitrailmod.blocks.WirePanel;
 
 public class ModBlocks {
 	public static Block closed_platform_top;
@@ -25,6 +27,8 @@ public class ModBlocks {
 	public static Block logo_block;
 	public static Block hung_arrow_sign;
 	public static Block platform_arrow_sign;
+	public static Block fluorescent_lamp;
+	public static Block wire_panel;
 	public static Block platform_sign;
 	
 	public static void init() {
@@ -36,6 +40,8 @@ public class ModBlocks {
 		logo_block = new LogoBlock(Material.iron);
 		hung_arrow_sign = new HungArrowSign(Material.iron);
 		platform_arrow_sign = new PlatformArrowSign(Material.iron);
+		fluorescent_lamp = new FluorescentLamp(Material.glass);
+		wire_panel = new WirePanel(Material.iron);
 		//platform_sign = new PlatformSign(Material.iron);
 	}
 	
@@ -48,6 +54,8 @@ public class ModBlocks {
 		GameRegistry.registerBlock(logo_block, logo_block.getUnlocalizedName().substring(5));
 		GameRegistry.registerBlock(hung_arrow_sign, hung_arrow_sign.getUnlocalizedName().substring(5));
 		GameRegistry.registerBlock(platform_arrow_sign, platform_arrow_sign.getUnlocalizedName().substring(5));
+		GameRegistry.registerBlock(fluorescent_lamp, fluorescent_lamp.getUnlocalizedName().substring(5));
+		GameRegistry.registerBlock(wire_panel, wire_panel.getUnlocalizedName().substring(5));
 		//GameRegistry.registerBlock(platform_sign, platform_sign.getUnlocalizedName().substring(5));
 	}
 	
@@ -59,18 +67,13 @@ public class ModBlocks {
 		registerRender(platform_panel_block);
 		registerRender(logo_block);
 		registerRender(hung_arrow_sign);
-		registerInventoryRender(hung_arrow_sign);
 		registerRender(platform_arrow_sign);
-		registerInventoryRender(platform_arrow_sign);
+		registerRender(fluorescent_lamp);
+		registerRender(wire_panel);
 		//registerRender(platform_sign);
 	}
 	
 	public static void registerRender(Block block){
-		Item item = Item.getItemFromBlock(block);
-		Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(item, 0, new ModelResourceLocation(References.MOD_ID + ":" + item.getUnlocalizedName().substring(5)));
-	}
-	
-	public static void registerInventoryRender(Block block){
 		Item item = Item.getItemFromBlock(block);
 		Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(item, 0, new ModelResourceLocation(References.MOD_ID + ":" + item.getUnlocalizedName().substring(5), "inventory"));
 	}