+++ src/main/java/tk/cth451/transitrailmod/init/ModRecipe.java
@@ -67,5 +67,22 @@ public class ModRecipe {
 		// platform arrow sign
 		GameRegistry.addShapelessRecipe(new ItemStack(ModBlocks.platform_arrow_sign, 2),
 				new Object[] {Blocks.iron_block, Items.arrow});
+		
+		// fluorescent lamp
+		GameRegistry.addRecipe(new ItemStack(ModBlocks.fluorescent_lamp, 16),
+				" A ",
+				" B ",
+				" C ",
+				'A', Blocks.glass,
+				'B', Blocks.glowstone,
+				'C', Blocks.iron_block);
+		
+		// fluorescent lamp
+		GameRegistry.addRecipe(new ItemStack(ModBlocks.wire_panel, 4),
+				"ABA",
+				"ABA",
+				"ABA",
+				'A', Items.iron_ingot,
+				'B', Items.redstone);
 	}
 }